+++ RELEASE_NOTES.md
@@ -1,14 +1,15 @@
-# Release 0.5.44
+# Release 0.5.45
 
 **Date:** 2026-04-29
 
 ## What's New
 
+- Merge feature 'auto-formulas-20260429-172854' into develop
 - Merge feature 'gitflow-auto-version-fallback' into develop
 - Squash(feature): makefile-wizard-logic-fix
-- Squash(feature): tap-publish-validation-automation
 
 ## Bug Fixes
 
+- Auto-backmerge before release finish
 - Fallback to next patch release version